src: Move projec code to lib
