+++ src/lib/cli.rs
