+++ src/lib/drivers/tcp/server.rs
@@ -1,14 +1,14 @@
 use std::sync::Arc;
 
 use anyhow::Result;
-use mavlink_server::callbacks::{Callbacks, MessageCallback};
 use tokio::{
-    net::TcpStream,
+    net::{TcpListener, TcpStream},
     sync::{broadcast, RwLock},
 };
 use tracing::*;
 
 use crate::{
+    callbacks::{Callbacks, MessageCallback},
     drivers::{
         tcp::{tcp_receive_task, tcp_send_task},
         Driver, DriverInfo,
@@ -20,8 +20,8 @@ use crate::{
     },
 };
 
-pub struct TcpClient {
-    pub remote_addr: String,
+pub struct TcpServer {
+    pub local_addr: String,
     name: arc_swap::ArcSwap<String>,
     uuid: DriverUuid,
     on_message_input: Callbacks<Arc<Protocol>>,
@@ -29,10 +29,10 @@ pub struct TcpClient {
     stats: Arc<RwLock<AccumulatedDriverStats>>,
 }
 
-pub struct TcpClientBuilder(TcpClient);
+pub struct TcpServerBuilder(TcpServer);
 
-impl TcpClientBuilder {
-    pub fn build(self) -> TcpClient {
+impl TcpServerBuilder {
+    pub fn build(self) -> TcpServer {
         self.0
     }
 
@@ -53,66 +53,92 @@ impl TcpClientBuilder {
     }
 }
 
-impl TcpClient {
+impl TcpServer {
     #[instrument(level = "debug")]
-    pub fn builder(name: &str, remote_addr: &str) -> TcpClientBuilder {
+    pub fn builder(name: &str, local_addr: &str) -> TcpServerBuilder {
         let name = Arc::new(name.to_string());
 
-        TcpClientBuilder(Self {
-            remote_addr: remote_addr.to_string(),
+        TcpServerBuilder(Self {
+            local_addr: local_addr.to_string(),
             name: arc_swap::ArcSwap::new(name.clone()),
-            uuid: Self::generate_uuid(remote_addr),
+            uuid: Self::generate_uuid(local_addr),
             on_message_input: Callbacks::new(),
             on_message_output: Callbacks::new(),
             stats: Arc::new(RwLock::new(AccumulatedDriverStats::new(
                 name,
-                &TcpClientInfo,
+                &TcpServerInfo,
             ))),
         })
+    }
+
+    /// Handles communication with a single client
+    #[instrument(
+        level = "debug",
+        skip(socket, hub_sender, on_message_input, on_message_output)
+    )]
+    async fn handle_client(
+        socket: TcpStream,
+        remote_addr: String,
+        hub_sender: Arc<broadcast::Sender<Arc<Protocol>>>,
+        on_message_input: Callbacks<Arc<Protocol>>,
+        on_message_output: Callbacks<Arc<Protocol>>,
+        stats: Arc<RwLock<AccumulatedDriverStats>>,
+    ) -> Result<()> {
+        let hub_receiver = hub_sender.subscribe();
+
+        let (read, write) = socket.into_split();
+
+        tokio::select! {
+            result = tcp_receive_task(read, &remote_addr, hub_sender, &on_message_input, &stats) => {
+                if let Err(e) = result {
+                    error!("Error in TCP receive task for {remote_addr}: {e:?}");
+                }
+            }
+            result = tcp_send_task(write, &remote_addr, hub_receiver, &on_message_output, &stats) => {
+                if let Err(e) = result {
+                    error!("Error in TCP send task for {remote_addr}: {e:?}");
+                }
+            }
+        }
+
+        debug!("Finished handling connection with {remote_addr}");
+        Ok(())
     }
 }
 
 #[async_trait::async_trait]
-impl Driver for TcpClient {
+impl Driver for TcpServer {
     #[instrument(level = "debug", skip(self, hub_sender))]
     async fn run(&self, hub_sender: broadcast::Sender<Arc<Protocol>>) -> Result<()> {
-        let server_addr = &self.remote_addr;
+        let listener = TcpListener::bind(&self.local_addr).await?;
         let hub_sender = Arc::new(hub_sender);
 
         loop {
-            debug!("Trying to connect to {server_addr:?}...");
-            let (read, write) = match TcpStream::connect(server_addr).await {
-                Ok(socket) => socket.into_split(),
+            match listener.accept().await {
+                Ok((socket, remote_addr)) => {
+                    let remote_addr = remote_addr.to_string();
+                    let hub_sender = hub_sender.clone();
+
+                    tokio::spawn(TcpServer::handle_client(
+                        socket,
+                        remote_addr,
+                        hub_sender,
+                        self.on_message_input.clone(),
+                        self.on_message_output.clone(),
+                        self.stats.clone(),
+                    ));
+                }
                 Err(error) => {
-                    error!("Failed connecting to {server_addr:?}: {error:?}");
+                    error!("Failed to accept TCP connection: {error:?}");
                     tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
-                    continue;
-                }
-            };
-            debug!("TcpClient successfully connected to {server_addr:?}");
-
-            let hub_receiver = hub_sender.subscribe();
-
-            tokio::select! {
-                result = tcp_receive_task(read, server_addr,  hub_sender.clone(), &self.on_message_input, &self.stats) => {
-                    if let Err(e) = result {
-                        error!("Error in TCP receive task: {e:?}");
-                    }
-                }
-                result = tcp_send_task(write, server_addr, hub_receiver, &self.on_message_output, &self.stats) => {
-                    if let Err(e) = result {
-                        error!("Error in TCP send task: {e:?}");
-                    }
                 }
             }
-
-            debug!("Restarting TCP Client connection loop...");
         }
     }
 
     #[instrument(level = "debug", skip(self))]
     fn info(&self) -> Box<dyn DriverInfo> {
-        return Box::new(TcpClientInfo);
+        return Box::new(TcpServerInfo);
     }
 
     fn name(&self) -> Arc<String> {
@@ -125,7 +151,7 @@ impl Driver for TcpClient {
 }
 
 #[async_trait::async_trait]
-impl AccumulatedDriverStatsProvider for TcpClient {
+impl AccumulatedDriverStatsProvider for TcpServer {
     async fn stats(&self) -> AccumulatedDriverStats {
         self.stats.read().await.clone()
     }
@@ -137,20 +163,21 @@ impl AccumulatedDriverStatsProvider for TcpClient {
     }
 }
 
-pub struct TcpClientInfo;
-impl DriverInfo for TcpClientInfo {
+pub struct TcpServerInfo;
+impl DriverInfo for TcpServerInfo {
     fn name(&self) -> &'static str {
-        "TcpClient"
+        "TcpServer"
     }
     fn valid_schemes(&self) -> &'static [&'static str] {
-        &["tcpc", "tcpclient"]
+        &["tcps", "tcpserver"]
     }
 
     fn cli_example_legacy(&self) -> Vec<String> {
         let first_schema = &self.valid_schemes()[0];
         let second_schema = &self.valid_schemes()[1];
         vec![
-            format!("{first_schema}:<IP>:<PORT>"),
+            format!("{first_schema}:<IP?>:<PORT>"),
+            format!("{first_schema}:14550"),
             format!("{first_schema}:0.0.0.0:14550"),
             format!("{second_schema}:127.0.0.1:14660"),
         ]
@@ -174,7 +201,7 @@ impl DriverInfo for TcpClientInfo {
         let host = url.host_str().unwrap();
         let port = url.port().unwrap();
         Some(Arc::new(
-            TcpClient::builder("TcpClient", &format!("{host}:{port}")).build(),
+            TcpServer::builder("TcpServer", &format!("{host}:{port}")).build(),
         ))
     }
 }
