+++ src/lib/drivers/udp/server.rs
@@ -1,7 +1,6 @@
 use std::{collections::HashMap, sync::Arc};
 
 use anyhow::Result;
-use mavlink_server::callbacks::{Callbacks, MessageCallback};
 use tokio::{
     net::UdpSocket,
     sync::{broadcast, RwLock},
@@ -9,6 +8,7 @@ use tokio::{
 use tracing::*;
 
 use crate::{
+    callbacks::{Callbacks, MessageCallback},
     drivers::{Driver, DriverInfo},
     protocol::{read_all_messages, Protocol},
     stats::{
