+++ src/lib/mod.rs
@@ -1,1 +1,8 @@
 pub mod callbacks;
+pub mod cli;
+pub mod drivers;
+pub mod hub;
+pub mod logger;
+pub mod protocol;
+pub mod stats;
+pub mod web;
