+++ src/lib/stats/accumulated/messages.rs
