+++ src/main.rs
@@ -1,13 +1,7 @@
-mod cli;
-mod drivers;
-mod hub;
-mod logger;
-mod protocol;
-mod stats;
-mod web;
-
 use anyhow::*;
 use tracing::*;
+
+use mavlink_server::{cli, hub, logger, web};
 
 #[tokio::main(flavor = "multi_thread", worker_threads = 10)]
 async fn main() -> Result<()> {
